+++ lib/matchers/validations/validations.rb
@@ -9,19 +9,20 @@ module Mongoid
           @type = validation_type.to_s
         end
 
+        def with_message(expected_message)
+          @expected_message = expected_message
+          self
+        end
+
         def matches?(subject)
-          @klass = class_of(subject)
+          @klass     = class_of(subject)
           @validator = detect_validator
+          @result    = true
 
-          if @validator
-            @negative_message = "#{@type.inspect} validator on #{@field.inspect}"
-            @positive_message = "#{@type.inspect} validator on #{@field.inspect}"
-          else
-            @negative_message = "no #{@type.inspect} validator on #{@field.inspect}"
-            return false
-          end
+          check_validator
+          check_message if @expected_message
 
-          true
+          @result
         end
 
         def failure_message
@@ -33,10 +34,31 @@ module Mongoid
         end
 
         def description
-          "validate #{@type.inspect} of #{@field.inspect}"
+          desc = "validate #{@type.inspect} of #{@field.inspect}"
+          desc << " with message: #{@expected_message.inspect}" if @expected_message
         end
 
         private
+
+        def check_validator
+          if @validator
+            @negative_message = "#{@type.inspect} validator on #{@field.inspect}"
+            @positive_message = "#{@type.inspect} validator on #{@field.inspect}"
+          else
+            @negative_message = "no #{@type.inspect} validator on #{@field.inspect}"
+            @result = false
+          end
+        end
+
+        def check_message
+          error_message = @validator.options[:message]
+          if @expected_message == error_message
+            @positive_message << " with message: #{error_message.inspect}"
+          else
+            @negative_message << " with message: #{error_message.inspect}"
+            @result = false
+          end
+        end
 
         def detect_validator
           @klass.validators_on(@field).detect { |v| v.kind.to_s == @type }