+++ test/models/models.rb
@@ -10,13 +10,14 @@ class Person
   field :age,      type: Integer
   field :email,    type: String
   field :role,     type: String
+  field :site,     type: String
 
   has_many :pets
 
   validates_presence_of(:name)
   validates_presence_of(:role, message: "no role")
   validates_uniqueness_of(:email)
-  validates_uniqueness_of(:login, case_sensitive: false)
+  validates_uniqueness_of(:login, scope: :site, case_sensitive: false)
   validates_length_of(:password, minimum: 8, maximum: 16)
   validates_format_of(:email, with: /^([^\s]+)((?:[-a-z0-9]\.)[a-z]{2,})$/i)
   validates_inclusion_of(:role, in: ["admin", "user"])