+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.0.2
 
++ Added `with_message` matcher for validation matchers to test custom messages.
 + Added `DocumentModuleMatcher` class to simplify document matchers.
 + Added `be_stored_in` document matcher.
 + Added `validate_inclusion_of` matcher.